Fixed problem with keys of length < 4
