+++ .gitignore
@@ -7,6 +7,7 @@
 
 #nuget specific files
 Packages
+.nuget
 
 # User-specific files
 *.suo
