+++ BinaryRage.UnitTests/BinaryRage.FunctionalTests.csproj
@@ -46,7 +46,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DBTests.cs" />
+    <Compile Include="Model.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SmallKeyTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
